+++ vira.hs
@@ -13,6 +13,7 @@
          , "./test/flake-parts" { overrideInputs = cf }
          , "./test/flake-parts-no-defaults" { overrideInputs = cf }
          , "./test/flake-parts-override-default" { overrideInputs = cf }
+         , "./test/flake-parts-override-default-cmd" { overrideInputs = cf }
          , "./test/standalone" { overrideInputs = cf }
          ]
      , signoff.enable = True